+++ scripts/2d/Main.js
@@ -1,10 +1,16 @@
+import {Color,Vec2} from './DataTypes.js';
 export class Main2d {//the main class is the canvas that get drawn on 
-    constructor(){
+    constructor(w=window.innerWidth,h=window.innerHeight,center=Vec2(0,0)){
         //this.c;
+        this.d=Vec2(w,h)
         this.canv=document.createElement("canvas");
         document.body.appendChild(this.canv);
-        this.canv.width=window.innerWidth;
-        this.canv.height=window.innerHeight;
+        this.canv.width=w;
+        this.canv.height=h;
+        this.canv.style.position='absolute';
+        this.canv.style.top=-center[1]+window.innerHeight/2 +'px';
+        this.canv.style.left=center[0]+window.innerWidth/2 +'px';
+        this.canv.style.transform='translate(-50%,-50%)';
         //
         document.body.style.overflow = 'hidden';
         document.body.style.position = 'fixed';
@@ -44,19 +50,18 @@ export class Main2d {//the main class is the canvas that get drawn on
         this.ctx.clearRect(0,0,window.innerWidth,window.innerHeight);
         while(true){
             if(this.funcs[0]===undefined){break;}
-            else if(performance.now()-cur>= this.funcs[0][1]){this.running.push(this.funcs.shift())}
+            else if(performance.now()-cur>= this.funcs[0][1]){this.imp(this.funcs.shift(),this.running)}
             else{break;}
         }
         for(const fn of this.running){
             fn[0](this.delta);
         }
-       for(let i=this.running.length-1;i>=0;i--){
-            if(performance.now()-cur>this.running[i][2]){
-                this.running.splice(i,1)
+       for(let i=0;i<this.running.length;i++){
+            if(performance.now()-cur>this.running[i][2] || this.running[i]?.[3]?.con){
+                this.running.splice(i,1);
+                i--;
             }
-           if(this.running[i]?.[3]?.con){
-               this.running.splice(i,1)
-           }
+           else{break;}
         }
     }
     refresh(){//this just calls process every frame (user should call it at the end of the animtion)
@@ -75,32 +80,32 @@ export class Main2d {//the main class is the canvas that get drawn on
         if(h<=60){
              r=(255-(1-s)*255)*l;
              g=(h*255/60+(1-s)*255)*l;
-            return [r,g,0,a];
+            return Color(r,g,0,a);
         }
         else if(h<=120){
             r=(255-(h-60)*255/60+(1-s)*255)*l;
             g=((255-(1-s)*255)*l);
-            return [r,g,0,a];
+            return Color(r,g,0,a);
         }
         else if(h<=180){
             g=(255-(1-s)*255)*l;
             b=((h-120)*255/60+(1-s)*255)*l;
-            return [0,g,b,a];
+            return Color(0,g,b,a);
         }
         else if (h<=180+60){
             b=(255-(1-s)*255)*l;
             g=(255-(h-180)*255/60+(1-s)*255)*l;
-            return [0,g,b,a];
+            return Color(0,g,b,a);
         }
         else if (h<=300){
             b=(255-(1-s)*255)*l;
             r=((h-240)*255/60+(1-s)*255)*l;
-            return [r,0,b,a];
+            return Color(r,0,b,a);
         }
         else{
             r=(255-(1-s)*255)*l;
             b=(255-(h-300)*255/60+(1-s)*255)*l;
-            return [r,0,b,a];
+            return Color(r,0,b,a);
         }
         
     }
@@ -180,5 +185,49 @@ export class Main2d {//the main class is the canvas that get drawn on
         const {main, ...obj}=entity;
         return JSON.stringify(obj);
     }
+    setCanv(w,h,center){
+        this.canv.style.top=-center[1]+window.innerHeight/2 +'px';
+        this.canv.style.left=center[0]+window.innerWidth/2 +'px';
+        this.d=Vec2(w,h);
+    }
 }
-
+/*
+export class Editor{
+    constructor(w,h,center,m=[10,10]){
+        this.editorDiv=document.createElement('div');
+        this.editorInp=document.createElement('textarea');
+        this.d=[w,h,center,m];
+    }
+    setup(){
+        this.editorDiv.style.position='absolute';
+        this.editorDiv.style.width=this.d[0]+this.d[3][0]+10+'px';
+        this.editorDiv.style.height=this.d[1]+this.d[3][1]+10+'px';
+        this.editorDiv.style.top=-this.d[2][1]+window.innerHeight/2+'px';
+        this.editorDiv.style.left=this.d[2][0]+window.innerWidth/2 +'px';
+        this.editorDiv.style.transform='translate(-50%,-50%)';
+        this.editorDiv.style.background=Color(100,100,100,1).getColor();
+        document.body.appendChild(this.editorDiv);
+        this.editorInp.style.position='absolute';
+        this.editorInp.style.width=this.d[0]+'px';
+        this.editorInp.style.height=this.d[1]+'px';
+        this.editorInp.style.top=-this.d[2][1]+window.innerHeight/2+'px';
+        this.editorInp.style.left=this.d[2][0]+window.innerWidth/2 +'px';
+        this.editorInp.style.transform='translate(-50%,-50%)';
+        document.body.appendChild(this.editorInp);
+        this.editorInp.addEventListener('keydown', (e) => {
+            if (e.key === 'Enter') {
+                e.preventDefault(); // stop newline if needed
+                const lines = this.editorInp.value.split('\n');
+                const lastLine = lines[lines.length - 1].trim();
+                this.handleCommand(lastLine); // custom function
+            }
+        });
+    }
+    handleCommand(lastLine){
+        console.log(lastLine);
+        this.editorInp.value=lastLine+'\n';
+        let val=eval(lastLine);
+        this.editorInp.value+=(val??'');
+        this.editorInp.value+='\n';
+    }
+}*/