+++ scripts/2d/PixelMod.js
@@ -10,7 +10,7 @@ export class PixelMod{
         this.Mod=null;
     }
     draw(){
-        this.Mod=this.main.ctx.getImageData((window.innerWidth-this.width)/2+this.center[0],(window.innerHeight-this.height)/2 -this.center[1],this.width,this.height);
+        this.Mod=this.main.ctx.getImageData((this.main.canv.width-this.width)/2+this.center[0],(this.main.canv.height-this.height)/2 -this.center[1],this.width,this.height);
         for(let y=0;y<this.height;y++){
             for(let x=0;x<this.width;x++){
                 this.loop(x,this.height-y-1);
@@ -20,6 +20,6 @@ export class PixelMod{
                 this.Mod.data[(y*this.width+x)*4+3]=this.currentPixelColor[3];
             }
         }
-        this.main.ctx.putImageData(this.Mod,(window.innerWidth-this.width)/2+this.center[0],(window.innerHeight-this.height)/2 -this.center[1])
+        this.main.ctx.putImageData(this.Mod,(this.main.canv.width-this.width)/2+this.center[0],(this.main.canv.height-this.height)/2 -this.center[1])
     }
 }