+++ scripts/2d/Vector.js
@@ -1,7 +1,8 @@
 import  {Color,Vec2,origin} from './DataTypes.js';
 import {Triangle} from './Triangle.js';
 export class Vector{
-    constructor({main,point1,point2,color=Color(0,0,0,1),thick=2,tipSize=3}){
+    constructor({main,point1=Vec2(0,0),point2=Vec2(100,100),color=Color(0,0,0,1),thick=2,tipSize=4}){
+        this.temp;
         this.point1=point1;
 	        this.point2=point2;  
         this.color=color;
@@ -16,19 +17,21 @@ export class Vector{
     draw(){
         let phi=Math.atan2(this.point2[1]-this.point1[1],this.point2[0]-this.point1[0]);
         let r =this.point2.sub(this.point1).getLength();// Math.sqrt((this.point2[1]-this.point1[1])**2+(this.point2[0]-this.point1[0])**2)
+       // console.log(this.main.canv.height-window.innerHeight);
         this.tip.point1=Vec2(r+this.point1[0],this.point1[1]);
         this.tip.point2=Vec2(r-this.tipSize+this.point1[0],this.tipSize+this.point1[1]);
         this.tip.point3=Vec2(r-this.tipSize+this.point1[0],-this.tipSize+this.point1[1]);
         this.tip.fillColor=this.color;
         this.main.ctx.beginPath();
-        this.main.ctx.moveTo(...this.point1.devFix())//this.point1[0]+window.innerWidth/2,-this.point1[1]+window.innerHeight/2);
+        this.main.ctx.moveTo(this.point1[0]+this.main.canv.width/2,-this.point1[1]+this.main.canv.height/2)//this.point1[0]+window.innerWidth/2,-this.point1[1]+window.innerHeight/2);
        // this.main.ctx.lineTo(this.point2[0]+window.innerWidth/2,-this.point2[1]+window.innerHeight/2);
-        this.main.ctx.lineTo(this.point1[0]+(r-this.tipSize)*Math.cos(phi)+window.innerWidth/2,-this.point1[1]-(r-this.tipSize)*Math.sin(phi)+window.innerHeight/2);
+        this.main.ctx.lineTo(this.point1[0]+(r-this.tipSize)*Math.cos(phi)+this.main.canv.width/2,-this.point1[1]-(r-this.tipSize)*Math.sin(phi)+this.main.canv.height/2);
         this.main.ctx.strokeStyle=this.color.getColor();
         this.main.ctx.lineWidth=this.thick;
         this.main.ctx.stroke();
         this.tip.theta=phi;
         this.tip.draw(Vec2(this.point1[0],this.point1[1]));
+        this.temp=Vec2(this.point1[0]+(r)*Math.cos(phi)+this.main.canv.width/2,-this.point1[1]-(r)*Math.sin(phi)+this.main.canv.height/2)
     }
     update(r,angle){
         //this.point2=this.point2.rot(this.point1,);
@@ -95,7 +98,13 @@ export class Vector{
             }
         }//
         this.main.addProcess([f,startTime,endTime]);
-    } 
-
+    }
+    bind(obj,t){
+        this.point1=obj.point1?obj.point1(t):this.point1;
+        this.point2=obj.point2?obj.point2(t):this.point2;
+        this.color=obj.color?obj.color(t):this.color;
+        this.tipSize=obj.tipSize?obj.tipSize(t):this.tipSize;
+        this.thick=obj.thick?obj.thick(t):this.thick;
+    }
 }